Moved Activites to the View package
