+++ app/src/main/res/layout/activitysearch.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchActivity">
+    tools:context=".Views.SearchActivity">
 
     <TextView
         android:id="@+id/dateTime2"
